# Fix function-url-quotes performance