--- a/.changeset/tender-apes-count.md
+++ b/.changeset/tender-apes-count.md
@@ -1,0 +1,5 @@
+---
+"stylelint": patch
+---
+
+Fixed: `function-url-quotes` performance