--- a/lib/rules/function-url-quotes/index.js
+++ b/lib/rules/function-url-quotes/index.js
@@ -57,6 +57,8 @@ const rule = (primary, secondaryOptions, context) => {
 		 * @param {import('postcss').Declaration} decl
 		 */
 		function checkDeclParams(decl) {
+			if (!/url\(/i.test(decl.value)) return;
+
 			if (!isStandardSyntaxDeclaration(decl)) return;
 
 			const value = getDeclarationValue(decl);