--- a/lib/rules/function-url-quotes/index.js
+++ b/lib/rules/function-url-quotes/index.js
@@ -24,6 +24,8 @@ const meta = {
 	fixable: true,
 };
 
+const URL_FUNC_REGEX = /url\(/i;
+
 /** @type {import('stylelint').Rule} */
 const rule = (primary, secondaryOptions, context) => {
 	return (root, result) => {
@@ -57,7 +59,7 @@ const rule = (primary, secondaryOptions, context) => {
 		 * @param {import('postcss').Declaration} decl
 		 */
 		function checkDeclParams(decl) {
-			if (!/url\(/i.test(decl.value)) return;
+			if (!URL_FUNC_REGEX.test(decl.value)) return;
 
 			if (!isStandardSyntaxDeclaration(decl)) return;