<!-- Each pull request must be associated with an open issue unless it's a documentation fix. If a corresponding issue does not exist, please create one so we can discuss the change first. -->

<!-- Please answer the following. We close pull requests that don't. -->

> Which issue, if any, is this issue related to?

Ref https://github.com/stylelint/stylelint/issues/6869

> Is there anything in the PR that needs further explanation?

Before:
Mean: 139.03539278571427 ms
Deviation: 19.870834517620747 ms

After:
Mean: 96.23857735714284 ms
Deviation: 12.178951548182841 ms